升级java17，整理代码。遗留问题：netty无法使用directbuffer

+++ src/java/net/minecraft/block/BlockBasePressurePlate.java
@@ -1,6 +1,5 @@
 package net.minecraft.block;
 
-import java.util.Random;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
@@ -12,36 +11,30 @@ import net.minecraft.util.EnumFacing;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
-public abstract class BlockBasePressurePlate extends Block
-{
-    protected BlockBasePressurePlate(Material materialIn)
-    {
+import java.util.Random;
+
+public abstract class BlockBasePressurePlate extends Block {
+    protected BlockBasePressurePlate(Material materialIn) {
         this(materialIn, materialIn.getMaterialMapColor());
     }
 
-    protected BlockBasePressurePlate(Material p_i46401_1_, MapColor p_i46401_2_)
-    {
+    protected BlockBasePressurePlate(Material p_i46401_1_, MapColor p_i46401_2_) {
         super(p_i46401_1_, p_i46401_2_);
         this.setCreativeTab(CreativeTabs.tabRedstone);
         this.setTickRandomly(true);
     }
 
-    public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos)
-    {
+    public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos) {
         this.setBlockBoundsBasedOnState0(worldIn.getBlockState(pos));
     }
 
-    protected void setBlockBoundsBasedOnState0(IBlockState state)
-    {
+    protected void setBlockBoundsBasedOnState0(IBlockState state) {
         boolean flag = this.getRedstoneStrength(state) > 0;
         float f = 0.0625F;
 
-        if (flag)
-        {
+        if (flag) {
             this.setBlockBounds(0.0625F, 0.0F, 0.0625F, 0.9375F, 0.03125F, 0.9375F);
-        }
-        else
-        {
+        } else {
             this.setBlockBounds(0.0625F, 0.0F, 0.0625F, 0.9375F, 0.0625F, 0.9375F);
         }
     }
@@ -49,79 +42,65 @@ public abstract class BlockBasePressurePlate extends Block
     /**
      * How many world ticks before ticking
      */
-    public int tickRate(World worldIn)
-    {
+    public int tickRate(World worldIn) {
         return 20;
     }
 
-    public AxisAlignedBB getCollisionBoundingBox(World worldIn, BlockPos pos, IBlockState state)
-    {
+    public AxisAlignedBB getCollisionBoundingBox(World worldIn, BlockPos pos, IBlockState state) {
         return null;
     }
 
     /**
      * Used to determine ambient occlusion and culling when rebuilding chunks for render
      */
-    public boolean isOpaqueCube()
-    {
+    public boolean isOpaqueCube() {
         return false;
     }
 
-    public boolean isFullCube()
-    {
+    public boolean isFullCube() {
         return false;
     }
 
-    public boolean isPassable(IBlockAccess worldIn, BlockPos pos)
-    {
+    public boolean isPassable(IBlockAccess worldIn, BlockPos pos) {
         return true;
     }
 
     /**
      * Return true if an entity can be spawned inside the block (used to get the player's bed spawn location)
      */
-    public boolean canSpawnInBlock()
-    {
+    public boolean canSpawnInBlock() {
         return true;
     }
 
-    public boolean canPlaceBlockAt(World worldIn, BlockPos pos)
-    {
+    public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
         return this.canBePlacedOn(worldIn, pos.down());
     }
 
     /**
      * Called when a neighboring block changes.
      */
-    public void onNeighborBlockChange(World worldIn, BlockPos pos, IBlockState state, Block neighborBlock)
-    {
-        if (!this.canBePlacedOn(worldIn, pos.down()))
-        {
+    public void onNeighborBlockChange(World worldIn, BlockPos pos, IBlockState state, Block neighborBlock) {
+        if (!this.canBePlacedOn(worldIn, pos.down())) {
             this.dropBlockAsItem(worldIn, pos, state, 0);
             worldIn.setBlockToAir(pos);
         }
     }
 
-    private boolean canBePlacedOn(World worldIn, BlockPos pos)
-    {
+    private boolean canBePlacedOn(World worldIn, BlockPos pos) {
         return World.doesBlockHaveSolidTopSurface(worldIn, pos) || worldIn.getBlockState(pos).getBlock() instanceof BlockFence;
     }
 
     /**
      * Called randomly when setTickRandomly is set to true (used by e.g. crops to grow, etc.)
      */
-    public void randomTick(World worldIn, BlockPos pos, IBlockState state, Random random)
-    {
+    public void randomTick(World worldIn, BlockPos pos, IBlockState state, Random random) {
     }
 
-    public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand)
-    {
-        if (!worldIn.isRemote)
-        {
+    public void updateTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
+        if (!worldIn.isRemote) {
             int i = this.getRedstoneStrength(state);
 
-            if (i > 0)
-            {
+            if (i > 0) {
                 this.updateState(worldIn, pos, state, i);
             }
         }
@@ -130,14 +109,11 @@ public abstract class BlockBasePressurePlate extends Block
     /**
      * Called When an Entity Collided with the Block
      */
-    public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn)
-    {
-        if (!worldIn.isRemote)
-        {
+    public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn) {
+        if (!worldIn.isRemote) {
             int i = this.getRedstoneStrength(state);
 
-            if (i == 0)
-            {
+            if (i == 0) {
                 this.updateState(worldIn, pos, state, i);
             }
         }
@@ -146,31 +122,25 @@ public abstract class BlockBasePressurePlate extends Block
     /**
      * Updates the pressure plate when stepped on
      */
-    protected void updateState(World worldIn, BlockPos pos, IBlockState state, int oldRedstoneStrength)
-    {
+    protected void updateState(World worldIn, BlockPos pos, IBlockState state, int oldRedstoneStrength) {
         int i = this.computeRedstoneStrength(worldIn, pos);
         boolean flag = oldRedstoneStrength > 0;
         boolean flag1 = i > 0;
 
-        if (oldRedstoneStrength != i)
-        {
+        if (oldRedstoneStrength != i) {
             state = this.setRedstoneStrength(state, i);
             worldIn.setBlockState(pos, state, 2);
             this.updateNeighbors(worldIn, pos);
             worldIn.markBlockRangeForRenderUpdate(pos, pos);
         }
 
-        if (!flag1 && flag)
-        {
-            worldIn.playSoundEffect((double)pos.getX() + 0.5D, (double)pos.getY() + 0.1D, (double)pos.getZ() + 0.5D, "random.click", 0.3F, 0.5F);
-        }
-        else if (flag1 && !flag)
-        {
-            worldIn.playSoundEffect((double)pos.getX() + 0.5D, (double)pos.getY() + 0.1D, (double)pos.getZ() + 0.5D, "random.click", 0.3F, 0.6F);
+        if (!flag1 && flag) {
+            worldIn.playSoundEffect((double) pos.getX() + 0.5D, (double) pos.getY() + 0.1D, (double) pos.getZ() + 0.5D, "random.click", 0.3F, 0.5F);
+        } else if (flag1 && !flag) {
+            worldIn.playSoundEffect((double) pos.getX() + 0.5D, (double) pos.getY() + 0.1D, (double) pos.getZ() + 0.5D, "random.click", 0.3F, 0.6F);
         }
 
-        if (flag1)
-        {
+        if (flag1) {
             worldIn.scheduleUpdate(pos, this, this.tickRate(worldIn));
         }
     }
@@ -178,16 +148,13 @@ public abstract class BlockBasePressurePlate extends Block
     /**
      * Returns the cubic AABB inset by 1/8 on all sides
      */
-    protected AxisAlignedBB getSensitiveAABB(BlockPos pos)
-    {
+    protected AxisAlignedBB getSensitiveAABB(BlockPos pos) {
         float f = 0.125F;
-        return new AxisAlignedBB((double)((float)pos.getX() + 0.125F), (double)pos.getY(), (double)((float)pos.getZ() + 0.125F), (double)((float)(pos.getX() + 1) - 0.125F), (double)pos.getY() + 0.25D, (double)((float)(pos.getZ() + 1) - 0.125F));
+        return new AxisAlignedBB((float) pos.getX() + 0.125F, pos.getY(), (float) pos.getZ() + 0.125F, (float) (pos.getX() + 1) - 0.125F, (double) pos.getY() + 0.25D, (float) (pos.getZ() + 1) - 0.125F);
     }
 
-    public void breakBlock(World worldIn, BlockPos pos, IBlockState state)
-    {
-        if (this.getRedstoneStrength(state) > 0)
-        {
+    public void breakBlock(World worldIn, BlockPos pos, IBlockState state) {
+        if (this.getRedstoneStrength(state) > 0) {
             this.updateNeighbors(worldIn, pos);
         }
 
@@ -197,43 +164,37 @@ public abstract class BlockBasePressurePlate extends Block
     /**
      * Notify block and block below of changes
      */
-    protected void updateNeighbors(World worldIn, BlockPos pos)
-    {
+    protected void updateNeighbors(World worldIn, BlockPos pos) {
         worldIn.notifyNeighborsOfStateChange(pos, this);
         worldIn.notifyNeighborsOfStateChange(pos.down(), this);
     }
 
-    public int getWeakPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side)
-    {
+    public int getWeakPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side) {
         return this.getRedstoneStrength(state);
     }
 
-    public int getStrongPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side)
-    {
+    public int getStrongPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side) {
         return side == EnumFacing.UP ? this.getRedstoneStrength(state) : 0;
     }
 
     /**
      * Can this block provide power. Only wire currently seems to have this change based on its state.
      */
-    public boolean canProvidePower()
-    {
+    public boolean canProvidePower() {
         return true;
     }
 
     /**
      * Sets the block's bounds for rendering it as an item
      */
-    public void setBlockBoundsForItemRender()
-    {
+    public void setBlockBoundsForItemRender() {
         float f = 0.5F;
         float f1 = 0.125F;
         float f2 = 0.5F;
         this.setBlockBounds(0.0F, 0.375F, 0.0F, 1.0F, 0.625F, 1.0F);
     }
 
-    public int getMobilityFlag()
-    {
+    public int getMobilityFlag() {
         return 1;
     }
