+++ src/java/net/minecraft/client/entity/EntityOtherPlayerMP.java
@@ -9,8 +9,7 @@ import net.minecraft.util.IChatComponent;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
 
-public class EntityOtherPlayerMP extends AbstractClientPlayer
-{
+public class EntityOtherPlayerMP extends AbstractClientPlayer {
     private boolean isItemInUse;
     private int otherPlayerMPPosRotationIncrements;
     private double otherPlayerMPX;
@@ -19,8 +18,7 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
     private double otherPlayerMPYaw;
     private double otherPlayerMPPitch;
 
-    public EntityOtherPlayerMP(World worldIn, GameProfile gameProfileIn)
-    {
+    public EntityOtherPlayerMP(World worldIn, GameProfile gameProfileIn) {
         super(worldIn, gameProfileIn);
         this.stepHeight = 0.0F;
         this.noClip = true;
@@ -31,26 +29,23 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
     /**
      * Called when the entity is attacked.
      */
-    public boolean attackEntityFrom(DamageSource source, float amount)
-    {
+    public boolean attackEntityFrom(DamageSource source, float amount) {
         return true;
     }
 
-    public void setPositionAndRotation2(double x, double y, double z, float yaw, float pitch, int posRotationIncrements, boolean p_180426_10_)
-    {
+    public void setPositionAndRotation2(double x, double y, double z, float yaw, float pitch, int posRotationIncrements, boolean p_180426_10_) {
         this.otherPlayerMPX = x;
         this.otherPlayerMPY = y;
         this.otherPlayerMPZ = z;
-        this.otherPlayerMPYaw = (double)yaw;
-        this.otherPlayerMPPitch = (double)pitch;
+        this.otherPlayerMPYaw = yaw;
+        this.otherPlayerMPPitch = pitch;
         this.otherPlayerMPPosRotationIncrements = posRotationIncrements;
     }
 
     /**
      * Called to update the entity's position/logic.
      */
-    public void onUpdate()
-    {
+    public void onUpdate() {
         this.renderOffsetY = 0.0F;
         super.onUpdate();
         this.prevLimbSwingAmount = this.limbSwingAmount;
@@ -58,22 +53,18 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
         double d1 = this.posZ - this.prevPosZ;
         float f = MathHelper.sqrt_double(d0 * d0 + d1 * d1) * 4.0F;
 
-        if (f > 1.0F)
-        {
+        if (f > 1.0F) {
             f = 1.0F;
         }
 
         this.limbSwingAmount += (f - this.limbSwingAmount) * 0.4F;
         this.limbSwing += this.limbSwingAmount;
 
-        if (!this.isItemInUse && this.isEating() && this.inventory.mainInventory[this.inventory.currentItem] != null)
-        {
+        if (!this.isItemInUse && this.isEating() && this.inventory.mainInventory[this.inventory.currentItem] != null) {
             ItemStack itemstack = this.inventory.mainInventory[this.inventory.currentItem];
             this.setItemInUse(this.inventory.mainInventory[this.inventory.currentItem], itemstack.getItem().getMaxItemUseDuration(itemstack));
             this.isItemInUse = true;
-        }
-        else if (this.isItemInUse && !this.isEating())
-        {
+        } else if (this.isItemInUse && !this.isEating()) {
             this.clearItemInUse();
             this.isItemInUse = false;
         }
@@ -83,27 +74,22 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
      * Called frequently so the entity can update its state every tick as required. For example, zombies and skeletons
      * use this to react to sunlight and start to burn.
      */
-    public void onLivingUpdate()
-    {
-        if (this.otherPlayerMPPosRotationIncrements > 0)
-        {
-            double d0 = this.posX + (this.otherPlayerMPX - this.posX) / (double)this.otherPlayerMPPosRotationIncrements;
-            double d1 = this.posY + (this.otherPlayerMPY - this.posY) / (double)this.otherPlayerMPPosRotationIncrements;
-            double d2 = this.posZ + (this.otherPlayerMPZ - this.posZ) / (double)this.otherPlayerMPPosRotationIncrements;
+    public void onLivingUpdate() {
+        if (this.otherPlayerMPPosRotationIncrements > 0) {
+            double d0 = this.posX + (this.otherPlayerMPX - this.posX) / (double) this.otherPlayerMPPosRotationIncrements;
+            double d1 = this.posY + (this.otherPlayerMPY - this.posY) / (double) this.otherPlayerMPPosRotationIncrements;
+            double d2 = this.posZ + (this.otherPlayerMPZ - this.posZ) / (double) this.otherPlayerMPPosRotationIncrements;
             double d3;
 
-            for (d3 = this.otherPlayerMPYaw - (double)this.rotationYaw; d3 < -180.0D; d3 += 360.0D)
-            {
-                ;
+            for (d3 = this.otherPlayerMPYaw - (double) this.rotationYaw; d3 < -180.0D; d3 += 360.0D) {
             }
 
-            while (d3 >= 180.0D)
-            {
+            while (d3 >= 180.0D) {
                 d3 -= 360.0D;
             }
 
-            this.rotationYaw = (float)((double)this.rotationYaw + d3 / (double)this.otherPlayerMPPosRotationIncrements);
-            this.rotationPitch = (float)((double)this.rotationPitch + (this.otherPlayerMPPitch - (double)this.rotationPitch) / (double)this.otherPlayerMPPosRotationIncrements);
+            this.rotationYaw = (float) ((double) this.rotationYaw + d3 / (double) this.otherPlayerMPPosRotationIncrements);
+            this.rotationPitch = (float) ((double) this.rotationPitch + (this.otherPlayerMPPitch - (double) this.rotationPitch) / (double) this.otherPlayerMPPosRotationIncrements);
             --this.otherPlayerMPPosRotationIncrements;
             this.setPosition(d0, d1, d2);
             this.setRotation(this.rotationYaw, this.rotationPitch);
@@ -112,20 +98,17 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
         this.prevCameraYaw = this.cameraYaw;
         this.updateArmSwingProgress();
         float f1 = MathHelper.sqrt_double(this.motionX * this.motionX + this.motionZ * this.motionZ);
-        float f = (float)Math.atan(-this.motionY * 0.20000000298023224D) * 15.0F;
+        float f = (float) Math.atan(-this.motionY * 0.20000000298023224D) * 15.0F;
 
-        if (f1 > 0.1F)
-        {
+        if (f1 > 0.1F) {
             f1 = 0.1F;
         }
 
-        if (!this.onGround || this.getHealth() <= 0.0F)
-        {
+        if (!this.onGround || this.getHealth() <= 0.0F) {
             f1 = 0.0F;
         }
 
-        if (this.onGround || this.getHealth() <= 0.0F)
-        {
+        if (this.onGround || this.getHealth() <= 0.0F) {
             f = 0.0F;
         }
 
@@ -136,14 +119,10 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
     /**
      * Sets the held item, or an armor slot. Slot 0 is held item. Slot 1-4 is armor. Params: Item, slot
      */
-    public void setCurrentItemOrArmor(int slotIn, ItemStack stack)
-    {
-        if (slotIn == 0)
-        {
+    public void setCurrentItemOrArmor(int slotIn, ItemStack stack) {
+        if (slotIn == 0) {
             this.inventory.mainInventory[this.inventory.currentItem] = stack;
-        }
-        else
-        {
+        } else {
             this.inventory.armorInventory[slotIn - 1] = stack;
         }
     }
@@ -151,16 +130,14 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
     /**
      * Send a chat message to the CommandSender
      */
-    public void addChatMessage(IChatComponent component)
-    {
+    public void addChatMessage(IChatComponent component) {
         Minecraft.getMinecraft().ingameGUI.getChatGUI().printChatMessage(component);
     }
 
     /**
      * Returns {@code true} if the CommandSender is allowed to execute the command, {@code false} if not
      */
-    public boolean canCommandSenderUseCommand(int permLevel, String commandName)
-    {
+    public boolean canCommandSenderUseCommand(int permLevel, String commandName) {
         return false;
     }
 
@@ -168,8 +145,7 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
      * Get the position in the world. <b>{@code null} is not allowed!</b> If you are not an entity in the world, return
      * the coordinates 0, 0, 0
      */
-    public BlockPos getPosition()
-    {
+    public BlockPos getPosition() {
         return new BlockPos(this.posX + 0.5D, this.posY + 0.5D, this.posZ + 0.5D);
     }
 }
