+++ src/java/net/minecraft/client/particle/EntityRainFX.java
@@ -8,10 +8,8 @@ import net.minecraft.util.BlockPos;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
 
-public class EntityRainFX extends EntityFX
-{
-    protected EntityRainFX(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn)
-    {
+public class EntityRainFX extends EntityFX {
+    protected EntityRainFX(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn) {
         super(worldIn, xCoordIn, yCoordIn, zCoordIn, 0.0D, 0.0D, 0.0D);
         this.motionX *= 0.30000001192092896D;
         this.motionY = Math.random() * 0.20000000298023224D + 0.10000000149011612D;
@@ -22,32 +20,28 @@ public class EntityRainFX extends EntityFX
         this.setParticleTextureIndex(19 + this.rand.nextInt(4));
         this.setSize(0.01F, 0.01F);
         this.particleGravity = 0.06F;
-        this.particleMaxAge = (int)(8.0D / (Math.random() * 0.8D + 0.2D));
+        this.particleMaxAge = (int) (8.0D / (Math.random() * 0.8D + 0.2D));
     }
 
     /**
      * Called to update the entity's position/logic.
      */
-    public void onUpdate()
-    {
+    public void onUpdate() {
         this.prevPosX = this.posX;
         this.prevPosY = this.posY;
         this.prevPosZ = this.posZ;
-        this.motionY -= (double)this.particleGravity;
+        this.motionY -= this.particleGravity;
         this.moveEntity(this.motionX, this.motionY, this.motionZ);
         this.motionX *= 0.9800000190734863D;
         this.motionY *= 0.9800000190734863D;
         this.motionZ *= 0.9800000190734863D;
 
-        if (this.particleMaxAge-- <= 0)
-        {
+        if (this.particleMaxAge-- <= 0) {
             this.setDead();
         }
 
-        if (this.onGround)
-        {
-            if (Math.random() < 0.5D)
-            {
+        if (this.onGround) {
+            if (Math.random() < 0.5D) {
                 this.setDead();
             }
 
@@ -61,32 +55,25 @@ public class EntityRainFX extends EntityFX
         block.setBlockBoundsBasedOnState(this.worldObj, blockpos);
         Material material = iblockstate.getBlock().getMaterial();
 
-        if (material.isLiquid() || material.isSolid())
-        {
+        if (material.isLiquid() || material.isSolid()) {
             double d0 = 0.0D;
 
-            if (iblockstate.getBlock() instanceof BlockLiquid)
-            {
-                d0 = (double)(1.0F - BlockLiquid.getLiquidHeightPercent(((Integer)iblockstate.getValue(BlockLiquid.LEVEL)).intValue()));
-            }
-            else
-            {
+            if (iblockstate.getBlock() instanceof BlockLiquid) {
+                d0 = 1.0F - BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL).intValue());
+            } else {
                 d0 = block.getBlockBoundsMaxY();
             }
 
-            double d1 = (double)MathHelper.floor_double(this.posY) + d0;
+            double d1 = (double) MathHelper.floor_double(this.posY) + d0;
 
-            if (this.posY < d1)
-            {
+            if (this.posY < d1) {
                 this.setDead();
             }
         }
     }
 
-    public static class Factory implements IParticleFactory
-    {
-        public EntityFX getEntityFX(int particleID, World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double xSpeedIn, double ySpeedIn, double zSpeedIn, int... p_178902_15_)
-        {
+    public static class Factory implements IParticleFactory {
+        public EntityFX getEntityFX(int particleID, World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double xSpeedIn, double ySpeedIn, double zSpeedIn, int... p_178902_15_) {
             return new EntityRainFX(worldIn, xCoordIn, yCoordIn, zCoordIn);
         }
     }
