+++ src/java/net/minecraft/client/renderer/block/model/BreakingFour.java
@@ -1,30 +1,26 @@
 package net.minecraft.client.renderer.block.model;
 
-import java.util.Arrays;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 
-public class BreakingFour extends BakedQuad
-{
+import java.util.Arrays;
+
+public class BreakingFour extends BakedQuad {
     private final TextureAtlasSprite texture;
 
-    public BreakingFour(BakedQuad quad, TextureAtlasSprite textureIn)
-    {
+    public BreakingFour(BakedQuad quad, TextureAtlasSprite textureIn) {
         super(Arrays.copyOf(quad.getVertexData(), quad.getVertexData().length), quad.tintIndex, FaceBakery.getFacingFromVertexData(quad.getVertexData()));
         this.texture = textureIn;
         this.remapQuad();
         this.fixVertexData();
     }
 
-    private void remapQuad()
-    {
-        for (int i = 0; i < 4; ++i)
-        {
+    private void remapQuad() {
+        for (int i = 0; i < 4; ++i) {
             this.remapVert(i);
         }
     }
 
-    private void remapVert(int vertex)
-    {
+    private void remapVert(int vertex) {
         int i = this.vertexData.length / 4;
         int j = i * vertex;
         float f = Float.intBitsToFloat(this.vertexData[j]);
@@ -33,8 +29,7 @@ public class BreakingFour extends BakedQuad
         float f3 = 0.0F;
         float f4 = 0.0F;
 
-        switch (this.face)
-        {
+        switch (this.face) {
             case DOWN:
                 f3 = f * 16.0F;
                 f4 = (1.0F - f2) * 16.0F;
@@ -65,7 +60,7 @@ public class BreakingFour extends BakedQuad
                 f4 = (1.0F - f1) * 16.0F;
         }
 
-        this.vertexData[j + 4] = Float.floatToRawIntBits(this.texture.getInterpolatedU((double)f3));
-        this.vertexData[j + 4 + 1] = Float.floatToRawIntBits(this.texture.getInterpolatedV((double)f4));
+        this.vertexData[j + 4] = Float.floatToRawIntBits(this.texture.getInterpolatedU(f3));
+        this.vertexData[j + 4 + 1] = Float.floatToRawIntBits(this.texture.getInterpolatedV(f4));
     }
 }
